requirements.txt
streamlit
openai
python-dotenv
plotly
pillow
